Selection of fast food restaurants in Toronto

A simple feature (sf) dataframe containing the locations of a selection of fast food restaurants in the city of Toronto, Canada (data are from 2008). The data are projected using EPSG: 26917 (WGS 84/UTM Zone 17N).

Usage

data(FastFood.sf)

Format

A simple features object with 614 rows and 3 variables:

ID

Unique identifier of record.

Class

Factor with 3 types of fast food restaurants: [P]izza, [S]andwich, and [H]amburger

geometry

Geometry of simple features.

Source

Ruiz et al. (2010) https://link.springer.com/article/10.1007/s10109-009-0100-1

References

Ruiz M, López FA, A Páez. (2010). Testing for spatial association of qualitative data using symbolic dynamics. Journal of Geographical Systems. 12 (3) 281-309