Testing for Spatial Structure of Qualitative Data in Cross Section.

spqdep offers the user a collection of functions to testing for spatial independence of categorical cross-section datasets.

Details

Some functionalities that have been included in spqdep package are:

1. Test Q

Testing for spatial dependence with the Q test based on the symbolic entropy. See references

2. Test QMap

Two statistics for the spatial analysis of qualitative data are obtained that are based on the symbolic entropy of the maps. See references

3. Spatial runs test

Testing for spatial dependence with the global test based on runs. See references

4. Local spatial runs test

Obtain the local test based on runs. See references

5. Scan test

Testing for spatial dependence with the Scan test. See references

6. Join-count tests

Testing for spatial dependence with the classical Join-count test. See references

7. Similarity test

Testing for spatial dependence with the Similarity test. See references

Datasets

(example provinces_spain, FastFood.sf) spqdep includes two different datasets: provinces_spain and FastFood.sf. These sets are used to illustrate the capabilities of different functions. Briefly, their main characteristics are the following

The FastFood.sf A sf object with points of the localization of Fastfood restaurants in Toronto.

The provinces_spain An object of the class sf with the geometry of spanish provinces and several data sets.

References

Ruiz M, López FA and A Páez (2011). Comparison of Thematic Maps Using Symbolic Entropy. International Journal of Geographical Information Science, 26, 413-439.

Ruiz, M., López, F., and Páez, A. (2010). Testing for spatial association of qualitative data using symbolic dynamics. Journal of Geographical Systems, 12(3), 281-309.0.

Kulldorff M, Nagarwalla N. (1995). Spatial disease clusters: Detection and Inference. Statistics in Medicine. 14:799-810

Jung I, Kulldorff M, Richard OJ (2010). A spatial scan statistic for multinomial data. Statistics in Medicine. 29(18), 1910-1918

Páez, A, López-Hernández, FA, Ortega-García, JA, & Ruiz, M. (2016). Clustering and co-occurrence of cancer types: A comparison of techniques with an application to pediatric cancer in Murcia, Spain. Spatial Analysis in Health Geography, 69-90.

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es