Summary of estimated objects of class spjctest.

This function summarizes estimated spjctest objects. The tables in the output include basic information for each test. blablabla...

Usage

# S3 method for class 'spjctest'
summary(object, ...)

Arguments

object

An spjctest object including a list of htest.

...

further arguments passed to or from other methods.

Value

An object of class summary.spjctest

See also

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

Examples

## Multinomial + Binomial using a sf multipolygon
data("provinces_spain")
sf::sf_use_s2(FALSE)
provinces_spain$Mal2Fml <- factor(provinces_spain$Mal2Fml > 100)
levels(provinces_spain$Mal2Fml) = c("men","woman")
provinces_spain$Older <- cut(provinces_spain$Older, breaks = c(-Inf,19,22.5,Inf))
levels(provinces_spain$Older) = c("low","middle","high")
f1 <- ~ Older + Mal2Fml
jc1 <- jc.test(formula = f1,
               data = provinces_spain,
               distr = "mc",
               alternative = "greater",
               zero.policy = TRUE)
#> although coordinates are longitude/latitude, st_intersects assumes that they
#> are planar
#> Warning: some observations have no neighbours;
#> if this seems unexpected, try increasing the snap argument.
#> Warning: neighbour object has 4 sub-graphs;
#> if this sub-graph count seems unexpected, try increasing the snap argument.
summary(jc1)

  

JoinCount Spatial Tests (Monte Carlo)
    

pairs
      	pvalue
      	Joincount
      	Expected
      	Variance
    

Older - multinomial - alternative: greater - Monte-Carlo simulation of
        join-count statistic (nonfree sampling)
    

high:high
	0.41600
	14
	13.92
	10.43

low:low
	0.59100
	7
	8.18
	6.04

middle:middle
	0.60500
	13
	13.79
	10.13

low:high
	0.45200
	24
	23.05
	16.26

middle:high
	0.65700
	28
	29.29
	17.47

middle:low
	0.27400
	25
	22.78
	15.38

Jtot
	0.37200
	77
	75.11
	20.17

Mal2Fml - binomial - alternative: greater - Monte-Carlo simulation of join-count statistic
    

men-men
	0.52400
	60
	60.37
	23.62

woman-woman
	0.02900
	13
	7.16
	6.66