A function to create symbols

This function obtains the set of symbols to get the Q-statistic

Usage

cr_symb(k = k, m = m)

Arguments

k

number of categories

m

length of the m-surrounding

Value

A list with two types of symbols. Permutation and Combinations-totals

p_symb	Matrix with symbols (permutations)

c_symb	Matrix with symbols (combinations)

Details

...

References

Ruiz M, López FA, A Páez. (2010). Testing for spatial association of qualitative data using symbolic dynamics. Journal of Geographical Systems. 12 (3) 281-309

See also

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

Examples

# Example 1: Obtain symbols for k=2 classes and m-surroundings of size 5
symb25 <- cr_symb(2,5)
symb25$p_symb # Permutations symbols
#>        [,1] [,2] [,3] [,4] [,5]
#> Perm1     1    1    1    1    1
#> Perm2     1    1    1    1    2
#> Perm3     1    1    1    2    1
#> Perm4     1    1    1    2    2
#> Perm5     1    1    2    1    1
#> Perm6     1    1    2    1    2
#> Perm7     1    1    2    2    1
#> Perm8     1    1    2    2    2
#> Perm9     1    2    1    1    1
#> Perm10    1    2    1    1    2
#> Perm11    1    2    1    2    1
#> Perm12    1    2    1    2    2
#> Perm13    1    2    2    1    1
#> Perm14    1    2    2    1    2
#> Perm15    1    2    2    2    1
#> Perm16    1    2    2    2    2
#> Perm17    2    1    1    1    1
#> Perm18    2    1    1    1    2
#> Perm19    2    1    1    2    1
#> Perm20    2    1    1    2    2
#> Perm21    2    1    2    1    1
#> Perm22    2    1    2    1    2
#> Perm23    2    1    2    2    1
#> Perm24    2    1    2    2    2
#> Perm25    2    2    1    1    1
#> Perm26    2    2    1    1    2
#> Perm27    2    2    1    2    1
#> Perm28    2    2    1    2    2
#> Perm29    2    2    2    1    1
#> Perm30    2    2    2    1    2
#> Perm31    2    2    2    2    1
#> Perm32    2    2    2    2    2
symb25$c_symb # Combinations-totals symbols
#>      level1 level2
#> [1,]      5      0
#> [2,]      4      1
#> [3,]      3      2
#> [4,]      2      3
#> [5,]      1      4
#> [6,]      0      5